💄 カテゴリページのSSG対応を強化し、データ取得ロジックを追加。HTMLサニタイズ機能をstringUtilsに移動し、コードを整理。
