+++ src/app/category/[categoryName]/[categoryNameId]/page.tsx
@@ -1,4 +1,4 @@
-export const runtime = "edge";
+// export const runtime = "edge";
 
 import "./page.css";
 import { Header } from "app/compornents/Header/Header";
@@ -6,6 +6,17 @@ import { Footer } from "app/compornents/Footer/Footer";
 import { Profile } from "app/compornents/profile/Profile";
 import Categoryblogs from "app/compornents/Categoryblogs/Categoryblogs";
 import { Blog } from "app/api/microcms/utils";
+import { generateStaticParams } from "app/lib/github/posts";
+import Search from "app/compornents/Search/Search";
+import CategoryPagination from "app/compornents/CategoryPagination/CategoryPagination";
+import { pagenationOffsetNum } from "app/utils/constants";
+import Maplist from "app/compornents/Maplist/Maplist";
+
+// SSGを強制
+export const dynamic = "force-static";
+
+// 更新間隔（秒）
+export const revalidate = 60; // 仮設定、最終は3600とする
 
 const BlogsCategoryName = async ({
   params,
@@ -50,6 +61,12 @@ const BlogsCategoryName = async ({
     return <div>データの読み込みに失敗しました</div>;
   }
 
+  // SSG用のデータ取得
+  const allPostsData = await generateStaticParams();
+
+  const pagenationOffset = pagenationOffsetNum; // 1ページあたりの表示件数
+  const totalPages = Math.ceil(allPostsData.length / pagenationOffset);
+
   const allBlogs: Blog[] = [
     ...data.contents,
     // ...(repoData
@@ -76,32 +93,51 @@ const BlogsCategoryName = async ({
   //   blog.tag?.some((blogTag) => blogTag.tag === tagId.tag)
   // );
   // console.log("getTagId", getTagId);
+  const uniqueTag = [{ id: "0", tag: categoryName }];
 
   return (
     <body>
       {/* <CustomHead /> */}
       <Header />
-      {/* <Maplist getTagId={getTagId} /> */}
-      <div id="container" className="flex w-4/5 h-auto m-auto mt-44">
-        <div id="main" className="w-full m-auto ml-4">
-          {/* Blog List */}
-          <h1 className="inline text-3xl font-bold pb-12"></h1>
+      <div className="top-container">
+        <div className="mt-44">
+          <Maplist getTagId={uniqueTag} />
+          {/* <Maplist getTagId={getTagId} /> */}
+        </div>
 
-          {/* 各投稿記事の表示 */}
-          <Categoryblogs
-            currentPage={currentPage}
+        <div id="container" className="flex w-11/12 h-auto m-auto">
+          <div
+            id="main"
+            className="grid grid-cols-2 sm:grid-cols-2 gap-y-8 w-full m-auto"
+          >
+            {/* Blog List */}
+            {/* <h1 className="inline text-3xl font-bold pb-12"></h1> */}
+
+            {/* 各投稿記事の表示 */}
+            <Categoryblogs
+              currentPage={currentPage}
+              categoryName={categoryName}
+              categoryNameId={categoryNameId}
+              fetchedData={allPostsData}
+              // fetchedData={data}
+            />
+          </div>
+
+          {/* プロフィール欄の表示 */}
+          <div className="flex flex-col ml-2">
+            <Search />
+            <Profile />
+          </div>
+        </div>
+        {/* ページ番号の記載 */}
+        <div className="mt-10 ml-12">
+          <CategoryPagination
+            totalPages={totalPages}
+            initialPage={currentPage}
             categoryName={categoryName}
-            categoryNameId={categoryNameId}
-            fetchedData={data}
           />
-
-          {/* ページ番号の記載 */}
-          {/* <Pagination totalPages={totalPages} initialPage={currentPage} /> */}
         </div>
-        {/* プロフィール欄の表示 */}
-        <div className="mt-8 ml-2">
-          <Profile />
-        </div>
+        {/* <Pagination totalPages={totalPages} initialPage={currentPage} /> */}
       </div>
       <Footer fetchedData={data} />
     </body>
